+++ src/zhinst/toolkit/exceptions.py
@@ -1,5 +1,17 @@
 """Global zhinst-toolkit exceptions."""
 
 
-class ValidationError(Exception):
-    """Data validation failed."""
+class ToolkitError(RuntimeError):
+    """Base class for `zhinst.toolkit` errors.
+
+    .. versionadded:: 0.5.2
+    """
+
+
+class ValidationError(ToolkitError):
+    """Data validation failed.
+
+    .. versionchanged:: 0.5.2
+
+         Changed base class from `Exception` to `ToolkitError`.
+    """